+++ packages/components/src/informative/ActivityAnswersBar/CustomLegend/CustomLegend.styles.js
@@ -1,10 +1,10 @@
 import { createStyles } from '@mantine/styles';
-import { pxToRem, getPaddings, getFontExpressive, getFontProductive } from '../../../theme.mixins';
 
 export const CustomLegendStyles = createStyles((theme, {}) => {
   return {
     root: {},
-    legendIcon: { color: theme.colors.mainWhite },
-    legendSlash: { fill: theme.colors.mainWhite, fontWeight: 600, fontSize: pxToRem(17) },
+    legendIconOK: { color: '#5CBC6A' },
+    legendIconKO: { color: '#D13B3B' },
+    legendIconSlash: { color: '#4D5358' },
   };
 });