+++ packages/components/src/overlay/Tooltip/Tooltip.styles.js
@@ -2,7 +2,7 @@ import { createStyles } from '@mantine/styles';
 import { pxToRem, getPaddings, getFontExpressive } from '../../theme.mixins';
 
 const getSizes = (theme, size) => {
-  const spacing = theme.spacing;
+  const { spacing } = theme;
 
   return {
     xs: {
@@ -24,17 +24,17 @@ const getSizes = (theme, size) => {
   }[size];
 };
 
-const getColor = (theme, color) => {
-  return {
+const getColor = (theme, color) =>
+  ({
     primary: {
-      backgroundColor: theme.colors.interactive01h,
+      backgroundColor: 'black', // theme.colors.interactive01h,
       color: theme.other.button.content.color.primary['default--reverse'],
       '&:hover': {
-        backgroundColor: theme.colors.interactive01h,
+        backgroundColor: 'black', // theme.colors.interactive01h,
       },
       '&:active': {
-        backgroundColor: theme.colors.interactive01h,
-        borderColor: theme.colors.interactive01h,
+        backgroundColor: 'black', // theme.colors.interactive01h,
+        // borderColor: theme.colors.interactive01h,
       },
     },
     secondary: {
@@ -48,26 +48,25 @@ const getColor = (theme, color) => {
         borderColor: theme.colors.interactive01,
       },
     },
-  }[color];
-};
+  })[color];
 
-export const TooltipStyles = createStyles((theme, { color, size }) => {
-  return {
-    root: {},
-    arrow: {
-      ...getColor(theme, color),
+const TooltipStyles = createStyles((theme, { color, size }) => ({
+  root: {},
+  arrow: {
+    ...getColor(theme, color),
+  },
+  tooltip: {
+    ...getFontExpressive(null, 400),
+    ...getColor(theme, color),
+    ...getSizes(theme, size),
+    display: 'flex',
+    flexDirection: 'row',
+    justifyContent: 'center',
+    alignItems: 'center',
+    svg: {
+      marginLeft: pxToRem(5),
     },
-    tooltip: {
-      ...getFontExpressive(null, 400),
-      ...getColor(theme, color),
-      ...getSizes(theme, size),
-      display: 'flex',
-      flexDirection: 'row',
-      justifyContent: 'center',
-      alignItems: 'center',
-      svg: {
-        marginLeft: pxToRem(5),
-      },
-    },
-  };
-});
+  },
+}));
+
+export { TooltipStyles };