+++ packages/leemons/src/common/AssetPlayer/components/ProgressBar/ProgressBar.js
@@ -1,15 +1,15 @@
 import React, { useEffect } from 'react';
-import { PROGRESS_BAR_DEFAULT_PROPS, PROGRESS_BAR_PROP_TYPES } from './ProgressBar.constants';
-import { ProgressBarStyles } from './ProgressBar.styles';
 import { Box, Text } from '@bubbles-ui/components';
 import {
-  ExpandFullIcon,
+  ExpandIcon,
   ControlsPlayIcon,
   ControlsPauseIcon,
   VolumeControlLowIcon,
   VolumeControlOffIcon,
 } from '@bubbles-ui/icons/solid';
 import { ShrinkIcon } from '@bubbles-ui/icons/outline';
+import { ProgressBarStyles } from './ProgressBar.styles';
+import { PROGRESS_BAR_DEFAULT_PROPS, PROGRESS_BAR_PROP_TYPES } from './ProgressBar.constants';
 
 const ProgressBar = ({
   seekValue,
@@ -47,25 +47,6 @@ const ProgressBar = ({
   const { classes } = ProgressBarStyles({}, { name: 'ProgressBar' });
   return (
     <Box className={classes.progressBarWrapper}>
-      <Box className={classes.progressBar}>
-        <Box
-          className={classes.progressBarValue}
-          style={{
-            width: `${playedPercentage}%`,
-          }}
-        />
-        <input
-          className={classes.progressBarSeekSlider}
-          type={'range'}
-          min={0}
-          max={0.999999}
-          step={'any'}
-          value={seekValue}
-          onChange={handleSeekChange}
-          onMouseDown={handleSeekMouseDown}
-          onMouseUp={handleSeekMouseUp}
-        />
-      </Box>
       <Box className={classes.controlBar}>
         <Box onClick={() => setIsPlaying(!isPlaying)} className={classes.iconWrapper}>
           {isPlaying ? (
@@ -94,10 +75,29 @@ const ProgressBar = ({
           style={{ flex: 1, justifyContent: 'right', marginRight: 2 }}
           onClick={handleFullscreenChange}
         >
+          <Box className={classes.progressBar}>
+            <Box
+              className={classes.progressBarValue}
+              style={{
+                width: `${playedPercentage}%`,
+              }}
+            />
+            <input
+              className={classes.progressBarSeekSlider}
+              type={'range'}
+              min={0}
+              max={0.999999}
+              step={'any'}
+              value={seekValue}
+              onChange={handleSeekChange}
+              onMouseDown={handleSeekMouseDown}
+              onMouseUp={handleSeekMouseUp}
+            />
+          </Box>
           {fullScreenMode ? (
             <ShrinkIcon height={20} width={20} className={classes.whiteIcon} />
           ) : (
-            <ExpandFullIcon height={20} width={20} className={classes.whiteIcon} />
+            <ExpandIcon height={20} width={20} className={classes.whiteIcon} />
           )}
         </Box>
       </Box>